Moves ProcessorSpec.get_all_connected_sets => QubitGraph.find_all_connected_sets.
This change was prompted by looking for get_* methods and noting that
this method of ProcessorSpec is currently unused within pyGSTi and
should really be a method of QubitGraph.
